Add WIP license
From Software Engineering StackExchange, user53019,
https://softwareengineering.stackexchange.com/a/277699 :

The MIT license doesn't specify attribution beyond maintaining
the copyright notice.

The above copyright notice and this permission notice shall be
included in all copies or substantial portions of the Software.
Somewhere within the source for the project you are considering
forking, there should be a LICENSE file or equivalent. If there
isn't, then you ought to create one.

The license file provides two key aspects:

Statement of copyright to the source.
License terms for use of the software.
To cover the first aspect, within the license file, you should
state something to this effect:

Copyright for portions of project Foo are held by [original author,
year] as part of project Bar. All other copyright for project Foo
are held by [your name, year].
And after that copyright statement, you should make sure the full
text for the MIT license is listed, which covers the second aspect.